refactor(auth): migrate authentication utility functions to a dedicated module and update imports across the application
