+++ frontend/src/features/auth/utils/authStorage.js
@@ -1,27 +1,3 @@
-const getApiBaseUrl = () => {
-  const url = import.meta.env.VITE_API_BASE_URL;
-
-  if (!url) {
-    throw new Error(
-      '❌ VITE_API_BASE_URL no está definida.\n' +
-        'Por favor, crea un archivo .env en la raíz del proyecto frontend con:\n' +
-        'VITE_API_BASE_URL=http://localhost:8080\n\n' +
-        'Para producción, configura la variable de entorno correspondiente.'
-    );
-  }
-
-  return url;
-};
-
-export const API_BASE_URL = getApiBaseUrl();
-
-export const API_TIMEOUT = 30000;
-
-export const DEFAULT_HEADERS = {
-  'Content-Type': 'application/json',
-  Accept: 'application/json'
-};
-
 const decodeJWT = token => {
   try {
     const base64Url = token.split('.')[1];
@@ -93,3 +69,4 @@ export const clearAuthData = () => {
   localStorage.removeItem('username');
   localStorage.removeItem('role');
 };
+
