+++ frontend/src/infrastructure/router/PublicRoute.jsx
@@ -1,13 +1,12 @@
 import { Navigate } from 'react-router-dom';
-import { getAuthToken } from '../constants/env.js';
+import { getAuthToken } from '@/features/auth/utils/authStorage.js';
 
 export const PublicRoute = ({ children }) => {
   const token = getAuthToken();
 
   if (token) {
-    return <Navigate to="/dashboard" replace />;
+    return <Navigate to='/dashboard' replace />;
   }
 
   return children;
 };
-
